Readability Refactor
refactor(data-plane): collapse observability submodule into single file with callback hook

Drops nemo_rl/data_plane/observability/ submodule (middleware.py 229 +
sinks.py 132 + __init__.py 34 = 395 LOC of MetricsSink ABC, InMemorySink,
LogSink, build_sink factory) in favor of one file
nemo_rl/data_plane/observability.py (160 LOC) with a single user-injected
on_event callback. The whole sink-ABC layer was speculative scaffolding
built before there was anything to observe — the lean shape preserves
per-op timing/transparency (the actual goal) while letting users plug
wandb / file / log via a function instead of subclassing a sink.

Public surface preserved: MetricsDataPlaneClient, print_event,
snapshot(). Factory wires the same way (cfg["observability"]["enabled"])
but takes cfg["observability"]["callback"] (programmatic) instead of
a sink Literal. Tests rewritten for the lean shape.

Net –232 LOC.


refactor(data-plane): move TQ awareness from BasePolicyWorker into TQWorkerMixin

The 14 TQ-aware methods (setup_data_plane, _fetch, _write_back,
train_presharded, get_logprobs_presharded, get_reference_policy_logprobs_presharded,
plus their helpers) and the _broadcast_batched_data_dict module-level
helper were appended to BasePolicyWorker so every worker carried TQ
awareness whether it used it or not. Lift them all into a separate
TQWorkerMixin under nemo_rl/data_plane/. Workers opt in by mixing it
into their MRO; bare workers stay zero-cost.

DTensor v1, V2, and Megatron worker subclasses inherit from
TQWorkerMixin AND AbstractPolicyWorker. The _get_replica_group overrides
those subclasses already had now satisfy the mixin's abstract hook.

base_policy_worker.py is now bit-identical to main (verified: git diff
main is empty for that file). All TQ awareness is reachable through
one file under nemo_rl/data_plane/.


refactor(policy): also extract _shard_for_logprob in get_topk_logits

The earlier shard-helper extraction collapsed the 27-line
dynamic-batches/sequence-packing/dense block in get_logprobs and
get_reference_policy_logprobs but missed get_topk_logits, which has
the same duplicated pattern. Apply the same extraction.

Pure refactor — same behavior. Companion sentinel change replaces
``self.use_dynamic_batches or self.use_sequence_packing`` with the
explicit ``unsorted_data_indices is not None``, matching the other
call sites.


fix(data-plane): make NoOp.kv_batch_get jagged-aware

The unconditional torch.stack at the bottom of kv_batch_get crashes on
per-token fields written via maybe_pack_jagged (variable row lengths).
Add _stack_or_nest helper: stack when shapes match, fall back to
torch.nested.as_nested_tensor when ragged. Mirrors what the TQ adapter
returns so codec.materialize takes the same branch on both adapters.

Net +18 LOC inside noop.py.


refactor(policy): move TQ-only result aggregators from lm_policy to tq_policy

The three aggregator helpers (_aggregate_train_results,
_aggregate_logprob_results, _aggregate_reference_logprob_results) were
defined at the top of lm_policy.py but called only by tq_policy.py's
*_from_meta methods. They have no caller on the legacy in-memory path.
Move them to tq_policy.py where they're used.

Pure code move — same behavior. lm_policy.py shrinks; tq_policy.py
absorbs the same lines. Eliminates the lm_policy → tq_policy
forward-reference import.


refactor(grpo): rename algorithms/sync_utils.py to algorithms/grpo_sync_workers.py

The file houses SyncRolloutActor (a Ray-remote rollout worker) plus
its kv_first_write helper, and is imported only by grpo_sync.py.
"sync_utils" is a generic name for what is, in practice, the worker
half of grpo_sync — rename for clarity, mirroring the existing
nemo_rl/algorithms/async_utils.py convention (one file containing the
async rollout actor + helpers).

Pure rename + import-string rewrite. No code change.


refactor(data-plane): tighten string-typed params with Literal types

Two stringly-typed parameters were validated at runtime but invisible
to static checkers:

- observability.py: _emit's status param accepts {"ok", "error",
  "timeout"}. Now typed as Literal[...] so callers see allowed values.
- worker_mixin.py: _fetch's layout ({"padded", "jagged"}) and
  fetch_policy ({"auto", "independent", "leader_broadcast"}) now
  Literal-typed for the same reason.

No behavior change.


refactor(experience): relocate SyncRolloutActor to experience/sync_rollout_actor.py

The earlier sync_utils.py → algorithms/grpo_sync_workers.py rename was
wrong-direction. SyncRolloutActor is a Ray-actor wrapper around the
stateless rollout building blocks in nemo_rl/experience/rollouts.py
(run_multi_turn_rollout, run_async_multi_turn_rollout,
run_async_nemo_gym_rollout). It belongs next to its dependency.

algorithms/ is for trainer orchestrators (grpo.py, grpo_sync.py);
data_plane/ is for the swappable client; the TQ write inside
SyncRolloutActor is incidental, not its primary identity. The actor
is a *colocated rollout worker* — domain match is experience/.

kv_first_write moves with the actor (one module, one file). Imports
updated in grpo_sync.py, ray_actor_environment_registry.py,
data_plane/preshard.py docstring, models/policy/tq_policy.py docstring,
and four test files.

No code change; pure relocation.
