+++ nemo_rl/data_plane/__init__.py
@@ -25,11 +25,14 @@ from nemo_rl.data_plane.interfaces import (
     DataPlaneConfig,
     KVBatchMeta,
 )
+from nemo_rl.data_plane.observability import MetricsDataPlaneClient, print_event
 
 __all__ = [
     "DataPlaneClient",
     "DataPlaneConfig",
     "KVBatchMeta",
+    "MetricsDataPlaneClient",
     "build_data_plane_client",
     "materialize",
+    "print_event",
 ]
