+++ nemo_rl/data_plane/preshard.py
@@ -80,7 +80,7 @@ def shard_meta_for_dp(
 
     Use this for every dispatch *after* rollout (logprob, ref-logprob, train).
     The rollout actor's first write is a flat ``kv_batch_put`` (see
-    :func:`nemo_rl.algorithms.sync_utils.kv_first_write`) — no fan-out.
+    :func:`nemo_rl.experience.sync_rollout_actor.kv_first_write`) — no fan-out.
 
     Per-rank packing metadata (``micro_batch_indices`` /
     ``micro_batch_lengths`` / ``elem_counts_per_gb``) lands in each shard's
