+++ nemo_rl/models/policy/workers/dtensor_policy_worker_v2.py
@@ -188,9 +188,12 @@ def get_train_context(
         yield
 
 
+from nemo_rl.data_plane.worker_mixin import TQWorkerMixin
+
+
 # Classes with @ray.remote can't be inherited from, so we split the implementation out.
 # This is useful when using worker extension classes.
-class DTensorPolicyWorkerV2Impl(AbstractPolicyWorker, ColocatablePolicyInterface):
+class DTensorPolicyWorkerV2Impl(TQWorkerMixin, AbstractPolicyWorker, ColocatablePolicyInterface):
     def __repr__(self) -> str:
         """Customizes the actor's prefix in the Ray logs.
