+++ tests/data_plane/unit/test_correctness.py
@@ -25,7 +25,7 @@ import pytest
 import torch
 from tensordict import TensorDict
 
-from nemo_rl.algorithms.sync_utils import kv_first_write
+from nemo_rl.experience.sync_rollout_actor import kv_first_write
 from nemo_rl.data_plane.adapters.noop import NoOpDataPlaneClient
 from nemo_rl.data_plane.driver_io import read_columns, write_columns
 from nemo_rl.data_plane.interfaces import KVBatchMeta
