+++ tests/data_plane/unit/test_observability.py
@@ -11,53 +11,52 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Unit tests for the data-plane observability middleware.
+"""Unit tests for the lean observability decorator.
 
-Uses :class:`NoOpDataPlaneClient` as the inner client so the tests run
-in the slim Tier-1 venv (no TQ, no Ray).
+Wraps :class:`NoOpDataPlaneClient` so the tests run in the slim Tier-1
+venv (no TQ, no Ray). The lean shape is one user-injected ``on_event``
+callback plus :meth:`snapshot` for cumulative totals — no ABC, no
+built-in sinks.
 """
 
 from __future__ import annotations
-
 
 import pytest
 import torch
 from tensordict import TensorDict
 
 from nemo_rl.data_plane.adapters.noop import NoOpDataPlaneClient
-from nemo_rl.data_plane.observability import (
-    InMemorySink,
-    MetricsDataPlaneClient,
-    build_sink,
-)
+from nemo_rl.data_plane.observability import MetricsDataPlaneClient
 
 
 @pytest.fixture
 def wrapped_client():
-    sink = InMemorySink()
+    events: list[dict] = []
     inner = NoOpDataPlaneClient()
-    yield MetricsDataPlaneClient(inner, sink=sink), sink
+    client = MetricsDataPlaneClient(inner, on_event=events.append)
+    yield client, events
     inner.close()
 
 
 def test_put_records_bytes_and_count(wrapped_client):
-    client, sink = wrapped_client
+    client, events = wrapped_client
     client.register_partition(
         partition_id="p", fields=["x"], num_samples=4, consumer_tasks=["read"]
     )
     fields = TensorDict({"x": torch.zeros(4, dtype=torch.float32)}, batch_size=[4])
     client.kv_batch_put(keys=["a", "b", "c", "d"], partition_id="p", fields=fields)
 
-    snap = sink.snapshot()
-    assert snap["data_plane/put/count"] == 1
-    # 4 floats * 4 bytes
-    assert snap["data_plane/put/bytes"] == 16
-    assert snap["data_plane/put/wall_ms"] >= 0
-    assert snap["data_plane/put/errors"] == 0
+    put_events = [e for e in events if e["op"] == "put"]
+    assert len(put_events) == 1
+    e = put_events[0]
+    assert e["status"] == "ok"
+    assert e["n_keys"] == 4
+    assert e["n_bytes"] == 16   # 4 floats * 4 bytes
+    assert e["wall_ms"] >= 0
 
 
 def test_get_records_after_put(wrapped_client):
-    client, sink = wrapped_client
+    client, events = wrapped_client
     client.register_partition(
         partition_id="p", fields=["x"], num_samples=2, consumer_tasks=["read"]
     )
@@ -68,36 +67,35 @@ def test_get_records_after_put(wrapped_client):
     out = client.kv_batch_get(keys=["a", "b"], partition_id="p", select_fields=["x"])
     assert torch.equal(out["x"], torch.ones(2))
 
-    snap = sink.snapshot()
-    assert snap["data_plane/get/count"] == 1
-    assert snap["data_plane/get/bytes"] > 0
+    get_events = [e for e in events if e["op"] == "get"]
+    assert len(get_events) == 1
+    assert get_events[0]["n_bytes"] > 0
 
 
 def test_register_and_clear_recorded(wrapped_client):
-    client, sink = wrapped_client
+    client, events = wrapped_client
     client.register_partition(
         partition_id="p", fields=["x"], num_samples=1, consumer_tasks=["r"]
     )
     client.kv_clear(keys=None, partition_id="p")
 
-    snap = sink.snapshot()
-    assert snap["data_plane/register/count"] == 1
-    assert snap["data_plane/clear/count"] == 1
+    ops = [e["op"] for e in events]
+    assert ops.count("register") == 1
+    assert ops.count("clear") == 1
 
 
-def test_error_counted_and_reraised(wrapped_client):
-    """Middleware does NOT swallow errors — re-raise after recording."""
-    client, sink = wrapped_client
-    # No register: kv_batch_get on an unknown partition should error.
+def test_error_status_recorded_and_reraised(wrapped_client):
+    """Decorator does NOT swallow errors — re-raise after recording."""
+    client, events = wrapped_client
     with pytest.raises(KeyError):
         client.kv_batch_get(keys=["a"], partition_id="nope", select_fields=["x"])
 
-    snap = sink.snapshot()
-    assert snap["data_plane/get/errors"] == 1
+    err = [e for e in events if e["op"] == "get" and e["status"] == "error"]
+    assert len(err) == 1
 
 
-def test_throughput_metric_emitted(wrapped_client):
-    client, sink = wrapped_client
+def test_snapshot_accumulates_successful_ops(wrapped_client):
+    client, _ = wrapped_client
     client.register_partition(
         partition_id="p", fields=["x"], num_samples=1, consumer_tasks=["r"]
     )
@@ -105,36 +103,36 @@ def test_throughput_metric_emitted(wrapped_client):
         keys=["a"], partition_id="p",
         fields=TensorDict({"x": torch.zeros(1)}, batch_size=[1]),
     )
-    snap = sink.snapshot()
-    assert "data_plane/put/throughput_MB_s" in snap
+    snap = client.snapshot()
+    assert snap["total_ops"] >= 2     # register + put
+    assert snap["total_bytes"] >= 4   # 1 float = 4 bytes
 
 
-def test_build_sink_factory():
-    assert isinstance(build_sink("memory"), InMemorySink)
-    assert isinstance(build_sink(None), InMemorySink)  # default
-    with pytest.raises(ValueError):
-        build_sink("not-a-real-sink")
+def test_default_callback_is_noop():
+    """Omitting on_event must not raise; the wrapper just forwards."""
+    inner = NoOpDataPlaneClient()
+    client = MetricsDataPlaneClient(inner)
+    client.register_partition(
+        partition_id="p", fields=["x"], num_samples=1, consumer_tasks=["r"]
+    )
+    client.close()
 
 
-def test_close_propagates_to_inner_and_sink(wrapped_client):
+def test_close_propagates(wrapped_client):
     client, _ = wrapped_client
     client.close()
-    # second close shouldn't raise
+    # Second close must not raise — NoOp is idempotent.
     client.close()
 
 
 def test_factory_wraps_when_observability_enabled():
-    """Factory + DataPlaneConfig integration — no real TQ needed."""
-    from nemo_rl.data_plane import build_data_plane_client
-
-    # Use NoOp impl path? Factory rejects 'noop'. Skip the real factory
-    # call and verify the wrap construction directly.
-    from nemo_rl.data_plane.adapters.noop import NoOpDataPlaneClient
-    from nemo_rl.data_plane.observability import (
-        InMemorySink,
-        MetricsDataPlaneClient,
-    )
-
-    client = MetricsDataPlaneClient(NoOpDataPlaneClient(), sink=InMemorySink())
+    """Programmatic wrap path; factory.py uses the same MetricsDataPlaneClient."""
+    inner = NoOpDataPlaneClient()
+    seen: list[dict] = []
+    client = MetricsDataPlaneClient(inner, on_event=seen.append)
     assert hasattr(client, "snapshot")
+    client.register_partition(
+        partition_id="p", fields=["x"], num_samples=1, consumer_tasks=["r"]
+    )
+    assert len(seen) == 1 and seen[0]["op"] == "register"
     client.close()
